requirements.txt
beautifulsoup4==4.10.0
python-dotenv==1.0.0
Requests==2.31.0
pyppeteer==2.0.0